+++ documentation/ChangeLog.md
@@ -6,3 +6,4 @@
    * Automatic indent/de-indent and brace completion on certain keypresses (Enter, Backspace). Configurable in `Preferences/Rust/Editor/Typing`.
  * Added: New Project wizard.
  * Added: Rust/Cargo project builder.
+   * With inline error marker reporting.